ref(ui) Move chart title components
I would like to use these in getsentry for the new subscription charts.
Importing from the views directories feels like bad hygiene. Moving
these components also lets us have fewer cross view imports as well.
