+++ src/sentry/static/sentry/app/views/performance/styles.tsx
@@ -17,26 +17,6 @@ export const DoubleHeaderContainer = styled('div')`
   grid-gap: ${space(3)};
 `;
 
-export const HeaderTitle = styled('div')`
-  display: inline-grid;
-  grid-auto-flow: column;
-  grid-gap: ${space(1)};
-  font-size: ${p => p.theme.fontSizeLarge};
-  color: ${p => p.theme.textColor};
-  align-items: center;
-`;
-
-export const HeaderTitleLegend = styled(HeaderTitle)`
-  background-color: ${p => p.theme.background};
-  border-bottom-right-radius: ${p => p.theme.borderRadius};
-  position: absolute;
-  z-index: 1;
-`;
-
-export const ChartContainer = styled('div')`
-  padding: ${space(2)} ${space(3)};
-`;
-
 export const ErrorPanel = styled('div')`
   display: flex;
   justify-content: center;
