+++ src/sentry/static/sentry/app/views/projectDetail/charts/projectBaseEventsChart.tsx
@@ -4,6 +4,7 @@ import {withTheme} from 'emotion-theming';
 
 import {fetchTotalCount} from 'app/actionCreators/events';
 import EventsChart from 'app/components/charts/eventsChart';
+import {HeaderTitleLegend} from 'app/components/charts/styles';
 import {getParams} from 'app/components/organizations/globalSelectionHeader/getParams';
 import {isSelectionEqual} from 'app/components/organizations/globalSelectionHeader/utils';
 import QuestionTooltip from 'app/components/questionTooltip';
@@ -13,7 +14,6 @@ import {axisLabelFormatter} from 'app/utils/discover/charts';
 import getDynamicText from 'app/utils/getDynamicText';
 import {Theme} from 'app/utils/theme';
 import withGlobalSelection from 'app/utils/withGlobalSelection';
-import {HeaderTitleLegend} from 'app/views/performance/styles';
 
 type Props = Omit<
   EventsChart['props'],
