+++ src/sentry/static/sentry/app/views/projectDetail/projectCharts.tsx
@@ -7,6 +7,7 @@ import {Location} from 'history';
 import {Client} from 'app/api';
 import OptionSelector from 'app/components/charts/optionSelector';
 import {
+  ChartContainer,
   ChartControls,
   InlineContainer,
   SectionHeading,
@@ -23,7 +24,6 @@ import {Theme} from 'app/utils/theme';
 import withApi from 'app/utils/withApi';
 
 import {getTermHelp, PERFORMANCE_TERM} from '../performance/data';
-import {ChartContainer} from '../performance/styles';
 
 import ProjectBaseEventsChart from './charts/projectBaseEventsChart';
 import ProjectStabilityChart from './charts/projectStabilityChart';
